+++ CHANGELOG.md
@@ -17,7 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ### Fixed
 - Updated `std` and `terser`.
 - Updated `react` to include typings.
-- `bundler` and `terser` plugins combinations.
+- Internal changes to how the processors and preprocessors are handled
+  to ensure that they are executed in the same order they were registered.
 
 ## [1.0.2] - 2021-08-20
 ### Added